refactor: move binFrequencies into NaChartCardBody component
The function was only used by this component. Colocating it removes
the module boundary violation (components/ importing agg/ internals).
Tests moved to tests/binFrequencies.test.ts.
